Add automatic network interface discovery

+++ lib/strings.c
@@ -18,7 +18,7 @@ String string_new(char *str) {
 void string_concat(String *dest, char *src) {
   int concat_len = strlen(src);
   
-  if (dest->capacity < concat_len + strlen(dest->buffer) + 1) { // null terminator
+  if ((long unsigned int)dest->capacity < concat_len + strlen(dest->buffer) + 1) { // null terminator
     dest->buffer = realloc(dest->buffer, (concat_len + dest->capacity));
   }
 
@@ -31,7 +31,7 @@ void string_concat(String *dest, char *src) {
 // SRC IS INVALID AFTER CALL, IF IT IS BUFFER FOR A STRING TYPE, FREE THE STRUCT AND MEMBERS
 void string_mutate(String *dest, char *src) {
   
-  if (dest->capacity < strlen(src) + 1) {
+  if ((long unsigned int)dest->capacity < strlen(src) + 1) {
     dest->buffer = realloc(dest->buffer, strlen(src) + 1);
   }
